backend/requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.4
django-enumchoicefield==1.0.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
nodeenv==1.3.0
pkg-resources==0.0.0
Pygments==2.2.0
PyJWT==1.6.1
pytz==2018.4
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
